refactoring: replaced validator-factory and validators into shared/forms.
