+++ src/modules/forms/types/index.ts
@@ -1,15 +1,4 @@
-import { UseControllerProps } from 'react-hook-form';
-
-export interface ControlledInputProps {
-  name: string;
-  label: string;
-  required?: boolean;
-  helperText?: string;
-  showError?: boolean;
-  rules?: UseControllerProps['rules'];
-}
-
 export interface LogInInputProps {
-  name: ControlledInputProps['name'];
+  name: string;
   disabled: boolean;
 }
