--- src/modules/forms/validation/constants/regExp.ts
@@ -1,6 +1,0 @@
-export const REGEXP_MATCH_UPPERCASE = /\p{Lu}/u;
-export const REGEXP_MATCH_DIGIT = /\d/;
-export const REGEXP_MATCH_SPECIAL_CHARACTER = /[!@#$%^&*(),.?":{}|<>]/;
-export const REGEXP_MATCH_ALL_WHITESPACES = /\s+/g;
-export const REGEXP_VALID_EMAIL =
-  /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
