--- src/modules/forms/validation/validators/requiredValidators.ts
@@ -1,7 +1,0 @@
-import createValidator from '../factory';
-
-const validateValueExists = (value: string) => {
-  return !!value;
-};
-
-export const fieldIsRequired = createValidator(validateValueExists);
