+++ src/shared/forms/index.ts
@@ -1,2 +1,4 @@
+export * from './model/validators';
+
 export { default as ControlledCheckbox } from './ui/ControlledCheckbox';
 export { default as ControlledTextField } from './ui/ControlledTextField';
