+++ src/shared/forms/model/regExp.ts
@@ -1,0 +1,5 @@
+export const MATCH_UPPERCASE = /\p{Lu}/u;
+export const MATCH_DIGIT = /\d/;
+export const MATCH_SPECIAL_CHARACTER = /[!@#$%^&*(),.?":{}|<>]/;
+export const MATCH_ALL_WHITESPACES = /\s+/g;
+export const VALID_EMAIL = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
